refactor(project) - scrum-21: change folders, add src
